WT-3824 Move tick calibration early - into global process structure.
