+++ src/support/time.c
@@ -81,6 +81,7 @@ __wt_tsc_to_nsec(WT_SESSION_IMPL *session, uint64_t end, uint64_t begin)
 {
 	double tsc_diff;
 
+	WT_UNUSED(session);
 	/*
 	 * If the ticks were reset, consider it an invalid check and just
 	 * return zero as the time difference because we cannot compute
@@ -89,7 +90,7 @@ __wt_tsc_to_nsec(WT_SESSION_IMPL *session, uint64_t end, uint64_t begin)
 	if (end < begin)
 		return (0);
 	tsc_diff = (double)(end - begin);
-	return ((uint64_t)(tsc_diff / S2C(session)->tsc_nsec_ratio));
+	return ((uint64_t)(tsc_diff / __wt_process.tsc_nsec_ratio));
 }
 
 /*
